checkpoint
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ibm@238972 138bc75d-0d04-0410-961f-82ee72b054a4

+++ stage-vector/gcc/testsuite/gcc.target/powerpc/vec-init-2.c
@@ -6,8 +6,8 @@
 #include <stddef.h>
 #include <altivec.h>
 
-extern void check (vector long a)       __attribute__((__noinline__));
-extern vector int pack (long a, long b) __attribute__((__noinline__));
+extern void check (vector long a)        __attribute__((__noinline__));
+extern vector long pack (long a, long b) __attribute__((__noinline__));
 
 void
 check (vector long a)
